Another refactor of the theme provider to add a layer of abstraction for the white label colors to be composed on top of a base theme.
