+++ layout/src/story-layout.tsx
@@ -1,11 +1,10 @@
 import React from 'react';
 import { ThemeProvider } from '../../theme-provider';
-import { themeList, Themes } from '../../theme-list/themeList';
 
 export const StoryLayout = {
   decorators: [
     (Story) => (
-      <ThemeProvider initialTheme={themeList[Themes.Light]}>
+      <ThemeProvider>
         <div style={{ padding: '1rem' }}>
           <Story />
         </div>
