--- theme-list/themes/bench-light.js
@@ -1,5 +1,0 @@
-import { baseTheme } from '../../theme/src/theme';
-
-export const BenchLight = {
-  ...baseTheme,
-};
