+++ theme-provider/types/whiteLabelColors.ts
@@ -1,0 +1,4 @@
+export type WhiteLabelColors = {
+  background?: string;
+  text?: string;
+};
