+++ theme/doc/borderWidths.mdx
@@ -1,8 +1,8 @@
 import { mdx } from '@mdx-js/react';
 import { MdxLayout } from '~/layout';
-import { baseTheme } from '../src/theme';
+import { lightBaseTheme } from '../src/theme';
 import BorderWidthsShowcase from './components/border-widths-showcase/BorderWidthsShowcase';
 export default MdxLayout;
 
 # Border Widths
-<BorderWidthsShowcase theme={baseTheme}/>
+<BorderWidthsShowcase theme={lightBaseTheme}/>
