+++ theme/doc/components/border-widths-sample/BorderWidthsSample.styles.tsx
@@ -4,7 +4,7 @@ export const Sample = styled('div', {
   width: 100,
   height: 30,
   background: '$surface-primary',
-  borderColor: '$onSurface-borderPrimary',
+  borderColor: '$onSurface-onPrimary',
   borderStyle: 'solid',
   borderRadius: '$radii$1',
 });
