+++ theme/doc/space.mdx
@@ -1,9 +1,9 @@
 import { mdx } from '@mdx-js/react';
 import { SystemUIShowcases } from '@divriots/dockit-react/systemui-showcases';
 import { MdxLayout } from '~/layout';
-  import { baseTheme } from '../src/theme';
+  import { lightBaseTheme } from '../src/theme';
 export default MdxLayout;
 
 # Space
 
-<SystemUIShowcases theme={baseTheme} showcaseKey="space" />
+<SystemUIShowcases theme={lightBaseTheme} showcaseKey="space" />
