+++ theme/src/themeBase.ts
@@ -1,0 +1,4 @@
+export enum ThemeBase {
+  Light = 'Light',
+  Dark = 'Dark',
+}
